+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Fix quaternion to Euler angles and Euler angles to quaternion conversion. The axes are now
+  correct, and the order the angles are applied is XYZ.
+
 ## [v0.9.1] - 2016-04-20
 
 ### Changed